# OpenAPI.yml
openapi: 3.0.0
servers:
  - description: Booking Manager Production
    url: https://www.booking-manager.com/api/v2
  - description: Booking Manager Beta
    url:  http://beta.booking-manager.com/api/v2
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/mmksystems/bm-api/2.0.0
info:
  description: |-
    # Booking Manager API for Agencies and Charter Fleets
    
    Booking Manager API is a standard Web service that enables clients to connect their data with external systems and services and to extend the automation benefits of the Booking Manager System.
    - It is used by charter agencies to publish availabilities of Charter Operators and automate booking processes by connecting Booking Manager to their own web sites to allow for seamless online booking experience for end users or integrating it with their CRM solutions in order to automate office procedures.
    - Charter operators use the web service to connect Booking Manager with their book keeping softwares and save time and effort in accounting department.
    - Third Party Software tools also use Booking Manager API to display richer information, for example boat tracking systems are able to visualise client contact information directly on the map showing the boat location in real time.
    ## Document Changes
    ### 29.05.2020
    - `/companies` more information added in response - address, additional telephone and additional mobile number
...
    ### 10.04.2020  
...
        
...
    
...
paths:
...
  '/companies':
    get:
      summary: Provides a list of companies
      operationId: getCompanies
      description: Calling */companies* retrieves a list of companies
      responses:
        '200':
          description: Retrieves the list of all the Charter operators in the Booking Manager System who manage their resources */yachts* availability in real time. This list provides the id's of the companies that are available for synchronisation and for real time booking. 
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
        '400':
          $ref: '#/components/responses/400'
        '401': 
          $ref: '#/components/responses/401'
      tags:
        - Booking
...
  '/yachts':
    get:
      summary: Provides a list of yachts
      operationId: getYachts
      description: Calling */yachts* retrieves a list of yachts, with all the detailed information and description of the boat. Description contains boat technical specifications, images and dedicated equipment categories. Also, information about applicable extra services for different products (Bareboat, Crewed, Cabin, Flotilla, Regatta and Allinclusive) are visible. One boat can have multiple products assigned (e.g. Bareboat and Crewed) at the same time, therefore each product has it s own elaboration of applicable extras.
      parameters:
        - $ref: '#/components/parameters/company'
      responses:
        '200':
          description: Explanation of the response structure <br /><br />
            First part  of the response contains technical info of the boat  <br /><br />
            id = unique resource id<br />
            name = the name of the individual yacht<br />
            model= name of the model<br />
            shipyardID= id of the shipyard (boat manufacturer)<br />
          
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Yacht'
        '400':
          $ref: '#/components/responses/400'
        '401': 
          $ref: '#/components/responses/401'
      tags:
        - Booking
...
  '/offers':
    get:
      summary: Provides a list of available offers
      operationId: getOffers
      description: Returns a list of all available yachts according to selected parameters such as -date from, date to, flexibility, product name, currency and trip duration. Flexibility presents timeframe (along with dates) as an additional filiter to retrieve results If you use flexibility 6 (in year), returned available boats will be only the ones that have default check in day Saturday. All other selected flexibility parameters will return all available departures (e.g Friday, Monday etc.)
      parameters:
        - $ref: '#/components/parameters/dateFrom'
        - $ref: '#/components/parameters/dateTo'
        - $ref: '#/components/parameters/flexibility'
        - $ref: '#/components/parameters/companyIdArray'
        - $ref: '#/components/parameters/countryArray'
        - $ref: '#/components/parameters/product'
        - $ref: '#/components/parameters/baseFromIdArray'
        - $ref: '#/components/parameters/baseToIdArray'
        - $ref: '#/components/parameters/sailingAreaIdArray'
        - $ref: '#/components/parameters/yachtIdArray'
        - $ref: '#/components/parameters/modelIdArray'
        - $ref: '#/components/parameters/currency'
        - $ref: '#/components/parameters/tripDurationArray'
        - $ref: '#/components/parameters/minCabins'
        - $ref: '#/components/parameters/maxCabins'
        - $ref: '#/components/parameters/minBerths'
        - $ref: '#/components/parameters/maxBerths'
        - $ref: '#/components/parameters/minHeads'
        - $ref: '#/components/parameters/maxHeads'
        - $ref: '#/components/parameters/minLength'
        - $ref: '#/components/parameters/maxLength'
      responses:
        '200':
          description: returns a list of yachts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Offer'
        '400':
          $ref: '#/components/responses/400'
        '401': 
          $ref: '#/components/responses/401'
        '422': 
          $ref: '#/components/responses/422'
      tags:
        - Booking
...
  '/reservation':
    post:
      summary: Create reservation
      operationId: createReservation
      description: Calling */reservation* is used to place a new option in the Booking Manager system. The system creates an option with standard discounts and obligatory applicable options. In this request, status is always 2 (option). 
      requestBody:
        description: json object
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Reservation'
      responses:
        '201':
          description: HTTPS status Created. Returns reservation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReservationResponse'
        '400':
          $ref: '#/components/responses/400'
        '401': 
          $ref: '#/components/responses/401'
        '422': 
          $ref: '#/components/responses/422'
      tags:
        - Booking
...
security:
  - bearerAuth: [] 
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    400:
      description: Bad request.
    401:
      description: Authorization Bearer token is missing or invalid.
...
    422:
      description: Unprocessable obligatory field.
  parameters:
...
    flexibility:
      in: query
      name: flexibility
      description: >
         `1` - on day `2` - in week `3` - one week `4` - two weeks `5` - in month `6` - in year
      required: false
      schema:
        $ref: "#/components/schemas/Flexibility"
    product:
      in: query
      name: productName
      description: product name
      required: false
      schema:
        $ref: "#/components/schemas/ProductEnum"
...
    dateFrom:
      in: query
      name: dateFrom
      description: date format&#58; yyyy-MM-ddTHH:mm:ss
      required: true
      schema:
         $ref: '#/components/schemas/DateTime'
    dateTo:
      in: query
      name: dateTo
      description: date format&#58; yyyy-MM-ddTHH:mm:ss
      required: true
      schema:
        $ref: '#/components/schemas/DateTime'
    companyIdArray:
      in: query
      name: companyId
      description: company ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a company id
          value: [191]
        multipleIds:
          summary: Example of multiple company ids
          value: [191, 724, 380]
    countryArray:
      in: query
      name: country
      description: countries
      required: false
      schema:
        type: array
        items:
          type: string
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a country id
          value: ["HR"]
        multipleIds:
          summary: Example of multiple country ids
          value: ["HR","IT"]
    baseFromIdArray:
      in: query
      name: baseFromId
      description: base from ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single base from id
          value: [123]
        multipleIds:
          summary: Example of multiple base from ids
          value: [123, 785, 859]
    baseToIdArray:
      in: query
      name: baseToId
      description: base to ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single base to id
          value: [123]
        multipleIds:
          summary: Example of multiple base to ids
          value: [123, 785, 859]
    sailingAreaIdArray:
      in: query
      name: sailingAreaId
      description: sailing area ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single sailing area id
          value: [15]
        multipleIds:
          summary: Example of multiple sailing area ids
          value: [4,15,22]
    yachtIdArray:
      in: query
      name: yachtId
      description: yacht ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single yacht id
          value: [123]
        multipleIds:
          summary: Example of multiple yachts ids
          value: [123, 785, 859]
    modelIdArray:
      in: query
      name: modelId
      description: model ids
      required: false
      schema:
        type: array
        items:
          $ref: '#/components/schemas/Long'
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single model id
          value: [123]
        multipleIds:
          summary: Example of multiple model ids
          value: [123, 785, 859]   
    currency:
      in: query
      name: currency
      description: currency
      required: false
      schema:
        type: string
      example: "EUR"
    tripDurationArray:
      in: query
      name: tripDuration
      description: one or more trip durations
      required: false
      schema:
        type: array
        items:
          type: integer
      style: form
      explode: true
      examples:
        oneId:
          summary: Example of a single trip duration
          value: [5]
        multipleIds:
          summary: Example of multiple trip durations
          value: [1, 5, 7]
    company:
      in: query
      name: companyId
      description: company id
      required: false
      schema:
         $ref: '#/components/schemas/Long'
    minCabins:
      in: query
      name: minCabins
      description: minimum number of cabins
      required: false
      schema:
        type: integer
        example: 5
    maxCabins:
      in: query
      name: maxCabins
      description: maximum number of cabins
      required: false
      schema:
        type: integer
        example: 10
    minBerths:
      in: query
      name: minBerths
      description: minimum number of berths
      required: false
      schema:
        type: integer
        example: 5
    maxBerths:
      in: query
      name: maxBerths
      description: maximum number of berths
      required: false
      schema:
        type: integer
        example: 10
    minHeads:
      in: query
      name: minHeads
      description: minimum number of heads
      required: false
      schema:
        type: integer
        example: 5
    maxHeads:
      in: query
      name: maxHeads
      description: maximum number of heads
      required: false
      schema:
        type: integer
        example: 10
    minLength:
      in: query
      name: minLength
      description: minimum length
      required: false
      schema:
        $ref: '#/components/schemas/Float'
    maxLength:
      in: query
      name: maxLength
      description: maximum length
      required: false
      schema:
        $ref: '#/components/schemas/Float'
  schemas:
    Long:
      type: integer
      format: int64
      example: 123456789
    Double:
      type: number
      format: double
      example: 6000.0
    Float:
      type: number
      format: float
      example: 12.51
    DateTime:
      type: string
      format: date-time
      example: "2019-01-01T00:00:00"
    Datetime:
      allOf:
        - $ref: '#/components/schemas/DateTime'
      example: "2019-01-01 00:00:00" 
    ProductEnum: 
      type: string
      enum: ["bareboat","crewed","cabin","flotilla","power","berth","regatta"]
    Flexibility:
      type: integer
      enum: [1,2,3,4,5,6]
...
    Company:
      type: object
      required: [id,name,city,zip,country,telephone,mobile,vatCode,email,web,bankAccountNumber,termsAndConditions]
      properties:
        id:
          $ref: '#/components/schemas/Long'
        name:
          type: string
          example: "DemoMMK"
        address:
          type: string
          example: "Radnička cesta 52"
        city:
          type: string
          example: "Zagreb"
        zip:
          type: string
          example: "123-45"
        country:
          type: string
          example: "Croatia"
        telephone:
          type: string
          example: "+3851 3773 951"
        telephone2:
          type: string
          example: "+3851 1234 567"
        mobile:
          type: string
          example: "+385 98/1234-567"
        vatCode:
          type: string
          example: "HR123456789"
        email:
          type: string
          example: "support@mmksystems.com"
        web:
          type: string
          example: "www.booking-manager.com"
        bankAccountNumber:
          type: string
          example: "HR123456789"
        termsAndConditions:
          type: string
          example: "Terms and conditions"
    Yacht:
      type: object
      required: [id,name,model,modelId,kind,homeBaseId,homebase,companyId,comapny]
      properties:
        id:
          type: integer
          format: int64
          example: 8860335000797
          description: unique resource id
        name:
          type: string
          example: "Bilko"
          description: the name of the individual yacht
        model:
          type: string
          example: "Sun Odyssey 509"
          description: the name of the individual yacht
        modelId:
          type: integer
          format: int64
          example: 428155400797
          description: the name of the model
        shipyardId:
          type: integer
          format: int64
          example: 5
          description: id of the shipyard (boat manufacturer)
        year:
          type: integer
          example: 2020
          description: the build year of the boat
        kind:
          type: string
          example: Gulet
          description: The kind of a vessel. Kind can be -  Sail boat , Motor boat, Catamaran, Power Catamaran, Gulet, Motorsailer, Motoryacht, Trimaran, Other
        homeBaseId: 
          type: integer
          format: int64
          example: 429492270000100000
          description: the unique id of the home base
        homeBase:
          type: string
          example: ACI Marina Split
          description: Home base name
        companyId: 
          type: integer
          format: int64
          example: 797
          description: ID of the charter company
        company:
          type: string
          example: "DemoMMK"
          description: Charter company name
        draught:
          type: number
          format: float
          example: 2.0
          description: draught (m)
        beam:
          type: number
          format: float
          example: 3.82
          description: beam (m)
        length:
          type: number
          format: float
          example: 12.8
          description: length at waterline (m)
        waterCapacity:
          type: number
          format: float
          example: 200.0
          description: capacity of the water tank (l)
        fuelCapacity:
          type: number
          format: float
          example: 170.0
          description: capacity of the fuel tank (l)
        engine:
          type: string
          example: "750 hp"
        deposit:
          type: number
          format: float
          example: 2000.0 
        commissionPercentage:
          type: number
          format: float
          example: 17.0
        wc:
          type: integer
          example: 2
        berths:
          type: integer
          example: 12
        cabins:
          type: integer
          example: 8
        mainsailArea:
          type: number
          format: float
          example: 75.34
        genoaArea:
          type: number
          format: float
          example: 28.0
        mainsailType:
          type: string
          example: "Furling"
          description: types - Full Batten or Furling
        genoaType:
          type: string
          example: "Self Tacking Jib"
          description : types - Furling or Self Tacking Jib
        defaultCheckInDay:
          type: integer
          example: 1
          description: -1 Any, 1 Sunday, 2 Monday, 3 Tuesday, 4 Wednesday, 5 Thursday, 6 Friday, 7 Saturday
        images:
          type: array
          items:
             $ref: '#/components/schemas/Image'
          description: Each boat has minimum of 3 pictures - Main, Interior and Plan Image. Additionally, each boat can have more pictures than 3 minimum ones, and if available, they will be visible along will all other pictures.
        equipmentIds:
          type: array
          items:
             type: integer
             format: int64
             example: 16
        products:
          type: array
          items: 
           $ref: '#/components/schemas/Product'
          description: Boat can have one or multiple products assigned on it. It means that it can be offered at the same time as one or more products ( e.g. Bareboat and Crewed). Values can be Bareboat, Crewed, Cabin, Flotilla, Regatta and Allinclusive.
         
        cabinsLayout:
          type: array
          items:
            $ref: '#/components/schemas/Layout'
        berthsLayout:
          type: array
          items:
            $ref: '#/components/schemas/Layout'
...
    Image:
      type: object
      required: [description,url,sortOrder]
      properties:
        description:
          type: string
          example: "Interior image"
        url:
          type: string
          example: "http://www.booking-manager.com/image123"
        sortOrder:
          type: integer
          example: 1
    Product:
      type: object
      required: [name, extras]
      properties:
        name:
          type: string
          example: "Bareboat"
        extras:
          type: array
          items:
            $ref: '#/components/schemas/Extras'
    Extras:
      type: object
      required: [name,obligatory,price,unit,payableInBase,includesDepositWaiver,validDaysFrom,validDaysTo]
      properties:
        name:
          type: string
          example: "Fuel"
        obligatory:
          type: boolean
          example: true
        price:
          type: number
          format: float
          example: 500.0
        unit:
          type: string
          example: "per booking"
        payableInBase:
          type: boolean
          example: true
        includedDepositWaiver:
          type: boolean
          example: false
        validDaysFrom:
          type: integer
          format: int64
          example: 0
        validDaysTo:
          type: integer
          format: int64
          example: 365
        minNumberOfPersons:
          type: integer
          format: int64
          example: 2
          description: only if > 0
        maxNumberOfPersons:
          type: integer
          format: int64
          example: 8
          description: only if > 0
        validForBases:
          type: array
          items:
            $ref: '#/components/schemas/ValidForBases'
    ValidForBases:
      type: object
      required: [from,to]
      properties:
        from:
          type: array
          items:
            type: integer
            format: int64
            example: 5
        to:
          type: array
          items:
            type: integer
            format: int64
            example: 27
    Layout: 
      type: object
      required: [name,amount]
      properties:
        name:
          type: string
          example: "1"
        amount:
          type: integer
          example: 8
    Offer:
      type: object
      required: [yachtId,yacht,startBaseId,endBaseId,startBase,endBase,dateFrom,dateTo,product,price,currency,startPrice,discountPercentage]
      properties:
        yachtId:
          $ref: '#/components/schemas/Long'
        yacht:
          type: string
          example: "Aurora"
        startBaseId:
          $ref: '#/components/schemas/Long'
        endBaseId:
          $ref: '#/components/schemas/Long'
        startBase:
          type: string
          example: "Dubrovnik"
        endBase:
          type: string
          example: "Dubrovnik"
        dateFrom:
          $ref: '#/components/schemas/Datetime'
        dateTo:
          $ref: '#/components/schemas/Datetime'
        product:
          type: string
          example: "Cabin"
        price:
          type: number
          format: double
          example: 5400
        currency:
          type: string
          example: "EUR"
        startPrice:
          $ref: '#/components/schemas/Double'
        discountPercentage:
          type: number
          format: float
          example: 10.00
...
    Reservation:
      type: object
      required: [dateFrom,dateTo,yachtId]
      properties:
        dateFrom:
          $ref: '#/components/schemas/DateTime'
        dateTo:
          $ref: '#/components/schemas/DateTime'
        yachtId:
          type: integer
          format: int64
          example: 978989630000100225
        status:
          type: integer
          format: int64
          example: 2
        productName:
          type: string
          example: "Bareboat"
        baseFromId:
          type: integer
          format: int64
          example: 667400400000100000
        baseToId:
          type: integer
          format: int64
          example: 667400400000100000
        clientName:
          type: string
          example: "John D."
        cliendId:
          $ref: '#/components/schemas/Long'
        currency:
          type: string
          example: "EUR"
    ReservationResponse:
      type: object
      required: [id,dateFrom,dateTo,yachtId,status,productName,baseFromId,baseToId,currency,clientName]
      properties:
        id:
          $ref: '#/components/schemas/Long'
        dateFrom:
          type: string
          format: date-time
          example: "2019-02-14T00:00:00"
        dateTo:
          type: string
          format: date-time
          example: "2019-02-21T00:00:00"
        expirationDate:
          type: string
          format: date-time
          example: "2019-02-12T00:00:00"
        yachtId:
          type: integer
          format: int64
          example: 978989630000100225
        status:
          $ref: '#/components/schemas/Long'
        productName:
          type: string
          example: "Bareboat"
        baseFromId:
          type: integer
          format: int64
          example: 667400400000100000
        baseToId:
          type: integer
          format: int64
          example: 667400400000100000'
        currency:
          type: string
          example: "EUR"
        clientName:
          type: string
          example: "John D."
...